Minimum wage dataset

A dataset containing number of people at each wage level in 25c bins, from 2007 to 2015. This is a subset of the dataset used in Roth and Sant'Anna (2023) that was in turn is derived from Cengiz et al. (2019). See those papers for additional descriptions.

Cengiz_df

Format

A data frame with 214812 rows and 9 variables:

wagebins

Wage bin

statenum

State ID

year

Year

quarterdate

Quarter

overallcountpc

Employment per capita in wage bin

treated_quarter

Was state treated that quarter

treated_year

Was state treated last 4 consecutive quarters

MW

Actual minimum wage

population

State population

Source

Roth and Sant'Anna (2023)

References

Cengiz, Doruk, Dube, Arindrajit, Lindner, Attila and Zipperer, Ben (2019). "The Effect of Minimum Wages on Low-Wage Jobs." The Quarterly Journal of Economics, vol. 134 (3), pp. 1405–1454, doi:10.1093/qje/qjz014

Roth, Jonathan and Sant'Anna, Pedro H. C. (2023), "When is Parallel Trends Sensitive to Functional Form?" Econometrica, vol. 91 (2), pp. 737–747, doi:10.3982/ECTA19402